Added filter-search box to interactive report. Placeholder file is left empty and kept. Some refactoring to move shared state properties to base report.
