--- testplan/web_ui/testing/src/Toolbar/FilterBoxPlaceholder.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-const FilterBoxPlaceholder = () => (
-  <div style={{ display: "flex", justifyContent: "center" }}>
-    <img
-      src="/testplan-logo.png"
-      alt=""
-      style={{
-        height: "1.5em",
-        marginRight: "0.2em",
-      }}
-    />
-    <b>Testplan</b>
-  </div>
-);
-
-export default FilterBoxPlaceholder;
